requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llvmlite==0.44.0
MarkupSafe==2.1.5
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
prettytable==3.11.0
pyecharts==2.0.6
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.6.1
scipy==1.15.1
simplejson==3.19.3
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
wcwidth==0.2.13
Werkzeug==3.1.3
openai>=1.0.0
python-dotenv>=1.0.0
